# static/specs/openapi/fe-gateway.yaml
openapi: 3.1.0
info:
  title: Eunomia DS Frontend Gateway API
  description: "OpenAPI specification for the Eunomia DS-Agent Frontend Gateway.

    The gateway proxies requests from the GUI to backend microservices.

    All paths are prefixed with `/api` and routed by service prefix to the

    corresponding microservice (catalog-agent, negotiation-agent, transfer-agent,
    auth-agent).

    "
  version: 1.0.0
servers:
  - url: "{gateway_url}/api"
    description: Frontend Gateway
    variables:
      gateway_url:
        default: http://localhost:8080
...
paths:
...
  /transfers/dataplane/dataplane-processes/{dataplaneProcessId}/events:
    get:
      tags:
        - Transfer Events
      summary: Get transfer events by dataplane process ID
      operationId: getTransferEventsByDataplaneProcessId
      parameters:
        - name: dataplaneProcessId
          in: path
          required: true
          schema:
            type: string
            format: urn
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TransferEventDto"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /events:
    get:
      tags:
        - Events
      summary: List events feed
      operationId: listEventsFeed
      parameters:
        - name: topic
          in: query
          required: false
          description: Topic pattern; `*` matches one segment and `**` any number of them.
          schema:
            type: string
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/PageNumber"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Sort"
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EventEnvelope"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      tags:
        - Events
      summary: Publish event
      operationId: publishEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishEventCommand"
      responses:
        "201":
          description: Event published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /events/{eventId}:
    get:
      tags:
        - Events
      summary: Get event by ID
      operationId: getEventById
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Event details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventEnvelope"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /events/stream:
    get:
      tags:
        - Events
      summary: Stream events via Server-Sent Events
      operationId: streamEventsSse
      description: EventSource cannot send headers, so the bearer token and the acting tenant go in the query.
      parameters:
        - name: topic
          in: query
          required: false
          schema:
            type: string
        - name: token
          in: query
          required: true
          schema:
            type: string
        - name: tenant
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: SSE event stream
          content:
            text/event-stream:
              schema:
                type: string
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /events/dlq:
    get:
      tags:
        - Events
      summary: List Dead Letter Queue records
      operationId: listDeadLetters
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [Unresolved, Replayed, Purged]
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/PageNumber"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Sort"
      responses:
        "200":
          description: List of dead letters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DeadLetterRecord"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /events/dlq/replay-all:
    post:
      tags:
        - Events
      summary: Replay all dead letter messages
      operationId: replayAllDeadLetters
      responses:
        "200":
          description: Number of messages replayed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReplayAllResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /events/dlq/{dlqId}:
    get:
      tags:
        - Events
      summary: Get Dead Letter Queue record
      operationId: getDeadLetterById
      parameters:
        - name: dlqId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Dead letter details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeadLetterRecord"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      tags:
        - Events
      summary: Delete Dead Letter Queue record
      operationId: deleteDeadLetter
      parameters:
        - name: dlqId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Dead letter deleted
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /events/dlq/{dlqId}/replay:
    post:
      tags:
        - Events
      summary: Replay a single dead letter message
      operationId: replayDeadLetter
      parameters:
        - name: dlqId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Delivery re-enqueued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventDeliveryRecord"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
components:
  parameters:
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        default: 10
        maximum: 100
    PageNumber:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
    Cursor:
      name: cursor
      in: query
      required: false
      description: Opaque cursor from `nextCursor` of the previous page; takes precedence over `page`.
      schema:
        type: string
    Sort:
      name: sort
      in: query
      required: false
      schema:
        type: string
        enum: [created_at_desc, created_at_asc, updated_at_desc, updated_at_asc]
        default: created_at_desc
...
  schemas:
...
    ErrorInfo:
      type: object
      required:
        - message
        - error_code
        - cause
      properties:
        message:
          type: string
        error_code:
          type: integer
        details:
          type: string
        cause:
          type: string
      examples:
        - message: Resource not found
          error_code: 3200
          status_code: 404
          cause: Catalog with ID urn:uuid:123 not found
...
    TransferEventDto:
      type: object
      properties:
        id:
          type: string
          format: urn
        transfer_id:
          type: string
          format: urn
        level:
          $ref: "#/components/schemas/LogLevel"
        component:
          type: string
        message:
          type: string
        data:
          type:
            - object
            - "null"
          additionalProperties: true
        created_at:
          type: string
          format: date-time
...
    LogLevel:
      type: string
      enum: [Debug, Info, Warn, Error]
...
    EventEnvelope:
      type: object
      required:
        - id
        - tenant_id
        - topic
        - source_crate
        - schema_version
        - timestamp
        - payload
      properties:
        id:
          type: string
          example: urn:uuid:f47ac10b-58cc-4372-a567-0e02b2c3d479
        tenant_id:
          type: string
        topic:
          type: string
          example: transfers.process.completed
        source_crate:
          type: string
          example: transfer-agent
        schema_version:
          type: integer
          example: 1
        correlation_id:
          type: string
          nullable: true
          example: urn:uuid:a1b2c3d4-e5f6-7890-abcd-ef1234567890
        timestamp:
          type: string
          format: date-time
          example: "2026-09-06T12:00:00Z"
        payload:
          type: object
          description: Any valid JSON event payload
          additionalProperties: true
    PublishEventCommand:
      type: object
      required:
        - topic
        - payload
      properties:
        topic:
          type: string
          example: custom.notification
        source_crate:
          type: string
          nullable: true
          example: gui-admin
        schema_version:
          type: integer
          nullable: true
          example: 1
        correlation_id:
          type: string
          nullable: true
          example: urn:uuid:a1b2c3d4-e5f6-7890-abcd-ef1234567890
        payload:
          type: object
          description: Event payload data
          additionalProperties: true
...
    EventDeliveryRecord:
      type: object
      required:
        - id
        - event_id
        - subscription_id
        - status
        - attempts
        - created_at
      properties:
        id:
          type: string
          example: del-123
        event_id:
          type: string
          example: urn:uuid:...
        subscription_id:
          type: string
          example: sub-456
        status:
          type: string
          enum: [Pending, Delivered, Failed, DeadLetter]
          example: Delivered
        attempts:
          type: integer
          example: 1
        last_attempt_at:
          type: string
          format: date-time
          nullable: true
        next_retry_at:
          type: string
          format: date-time
          nullable: true
        error_message:
          type: string
          nullable: true
        response_status_code:
          type: integer
          nullable: true
          example: 200
        delivered_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
    DeadLetterRecord:
      type: object
      required:
        - id
        - event_id
        - subscription_id
        - topic
        - callback_address
        - payload
        - error_message
        - attempts
        - status
        - failed_at
      properties:
        id:
          type: string
          example: dlq-123
        delivery_id:
          type: string
          nullable: true
          example: del-456
        event_id:
          type: string
          example: urn:uuid:...
        subscription_id:
          type: string
          example: sub-789
        topic:
          type: string
          example: transfers.started
        callback_address:
          type: string
          example: https://unreachable.host.com/webhook
        payload:
          type: object
          additionalProperties: true
        error_message:
          type: string
          example: "Connection refused: connect"
        attempts:
          type: integer
          example: 5
        status:
          type: string
          enum: [Unresolved, Replayed, Purged]
          example: Unresolved
        failed_at:
          type: string
          format: date-time
        replayed_at:
          type: string
          format: date-time
          nullable: true
    ReplayAllResponse:
      type: object
      required:
        - replayed_count
      properties:
        replayed_count:
          type: integer
          example: 3
  responses:
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorInfo"
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorInfo"
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorInfo"
...